Trying to fix CI

+++ clean.sh
@@ -4,4 +4,6 @@ SOURCE_FILE="src/TemperatureHumidityDHTSensorFan/TemperatureHumidityDHTSensorFan
 
 sed -i "s/#define VERSION .*/#define VERSION \"1-0-0-0\"/" $SOURCE_FILE
 
-rm tests/nunit/src/TemperatureHumidityDHTSensorFan.Tests.Integration/obj -R
+if [ -d tests/nunit/src/TemperatureHumidityDHTSensorFan.Tests.Integration/obj ]; then
+  rm tests/nunit/src/TemperatureHumidityDHTSensorFan.Tests.Integration/obj -R
+fi
